fix todos

+++ done/pictDBM/db_delete.c
@@ -55,7 +55,8 @@ int do_delete (const char* pict_id, struct pictdb_file* db_file)
 
     // By default fwrite start writing at the begining of the stream. Therefore, we
     // need to move the cursor to the position of where we want to delete
-    if (fseek(db_file->fpdb, sizeof(struct pictdb_header) + pict_delete_offset * sizeof(struct pict_metadata), SEEK_SET)
+    if (fseek(db_file->fpdb, (long) (sizeof(struct pictdb_header) + pict_delete_offset * sizeof(struct pict_metadata)),
+              SEEK_SET)
         != 0 ||
         fwrite(pict_to_delete, sizeof(struct pict_metadata), 1, db_file->fpdb) != 1) {
         return ERR_IO;
